Fixes a path issue with when the loc module loads
If you change the Primary Localization Directory path using the editor drop down, it will append /Game/* to the front. So for the stats window, that is expected as it gives us /Game/*/loc_en.csv. However, when loading in the string table, it will also use that path resulting in /Content//Game/*/loc_en.csv. This fix removes any project paths from the file name because Internal_LocTableFromFile factors the project path in.